refactor: rename `io` package to `view`
