+++ src/main/java/racingcar/view/InputMessages.java
@@ -1,4 +1,4 @@
-package racingcar.io;
+package racingcar.view;
 
 public class InputMessages {
