+++ src/main/java/racingcar/view/Output.java
@@ -1,4 +1,4 @@
-package racingcar.io;
+package racingcar.view;
 
 public interface Output {
