+++ src/test/java/racingcar/view/KeyboardInputParserTest.java
@@ -1,4 +1,4 @@
-package racingcar.io;
+package racingcar.view;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
