+++ src/test/java/racingcar/view/KeyboardInputTest.java
@@ -1,4 +1,4 @@
-package racingcar.io;
+package racingcar.view;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
